pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
